+++ R/cre.R
@@ -48,7 +48,7 @@
 #'  - *replace*: Boolean variable for replacement in bootstrapping.
 #'  - *max_decay*: Decay Threshold for pruning the rules.
 #'  - *type_decay*: Decay Type for pruning the rules (1: relative error; 2: error).
-#'  - *t_anom*: The threshold to define too generic or too specific (anomalous) rules.
+#'  - *t_ext*: The threshold to define too generic or too specific (extreme) rules.
 #'  - *t_corr*: The threshold to define correlated rules.
 #'  - *stability_selection*: Whether or not using stability selection for
 #'  selecting the causal rules.
@@ -97,7 +97,7 @@
 #'                      max_depth = 15,
 #'                      max_decay = 0.025,
 #'                      type_decay = 2,
-#'                      t_anom = 0.025,
+#'                      t_ext = 0.025,
 #'                      t_corr = 1,
 #'                      replace = FALSE,
 #'                      stability_selection = TRUE,