+++ tests/testthat/test-generate_causal_rules.R
@@ -32,7 +32,7 @@ test_that("generate_causal_rules works as expected!", {
                       max_depth = 15,
                       max_decay = 0.025,
                       type_decay = 2,
-                      t_anom = 0.025,
+                      t_ext = 0.025,
                       t_corr = 1,
                       replace = FALSE,
                       stability_selection = TRUE,